Refactor into modules just a bit
