+++ src/util.rs
@@ -1,0 +1,30 @@
+//! General utilities that have nothing to do with lattices
+
+
+#[derive(Debug, Clone, Copy)]
+pub enum Axis {
+    X,
+    Y,
+    Z,
+}
+
+
+impl Axis {
+    pub fn map<'a>(prefix: Option<String>) -> Vec<(String, Axis)> {
+        let axes = vec![("x", Axis::X), ("y", Axis::Y), ("z", Axis::Z)];
+        match prefix {
+            Some(p) => axes.into_iter().map(|(k, i)| (format!("{}{}", p, k), i)).collect(),
+            None => axes.into_iter().map(|(k, i)| (k.to_string(), i)).collect(),
+        }
+    }
+}
+
+
+
+pub fn python_mod(num: i32, modulus: usize) -> (i32, i32) {
+    if num < 0 {
+        (modulus as i32 + num % modulus as i32, num / modulus as i32 - 1)
+    } else {
+        (num % modulus as i32, num / modulus as i32)
+    }
+}
